tests: split WebRoutesTests.swift (1,262 lines → 5 files)
Follow-on to #490.

Layout:
  WebRoutesTestCase.swift            — base + helpers (auth, seeding,
                                       multipart, submitOnceAs)
  WebRoutesIndexTests.swift          — GET / dashboard (10)
  WebRoutesSubmitHistoryTests.swift  — submit form + history (5)
  WebRoutesSubmissionPageTests.swift — result page + tier visibility (18)
  WebRoutesBadgeTests.swift          — v0.4.127 class-wide badge gating (3)

36 tests preserved; all pass under --parallel.
